+++ scripts/aggregate-translations/aggregate-translations.js
@@ -2,7 +2,7 @@ const path = require('path');
 const fse = require('fs-extra');
 const glob = require('glob');
 const supportedLocales = require('./i18nSupportedLocales');
-const Logger = require('../../lib/logger');
+const Logger = require('../utils/logger');
 
 const aggregateMessages = require('./aggregate-messages');
 const writeAggregatedTranslations = require('./write-aggregated-translations');
